DATA: Added sql scripts for tables creation

+++ WheelCoin.Domain/Entites/Wheel.cs
@@ -6,7 +6,6 @@ public class Wheel : Auditable
 {
     public Attachment Image { get; set; }
     public string Name { get; set; }
-
-    public int Price { get; set; }
-    public int IncomeAmount { get; set; }
+    public long Price { get; set; }
+    public long IncomeAmount { get; set; }
 }
